requirements.txt
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7
Pillow==10.4.0
numpy==2.1.3
